Add Group Weapons & Gadgets
This will be used to hold mods that need to be placed after sorting mods. As sorting mods cover all vanilla items, any mods alter vanilla items or adding new items should be placed after the sorting mods.
Patches for sorting mods generally do not have the sorting mods as a master, so any mod that requires a patch should be placed after the sorting mod.